+++ ChangeLog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 1.2.1 (Unreleased)
+## 1.3.0
 
 Release notes: